fixing pod_test.go restruct of pod.go
