+++ pkg/kubernetes/client.go
@@ -14,7 +14,7 @@ import (
 	"k8s.io/client-go/tools/clientcmd"
 )
 
-// TODO: this moved to unstructured pkg, check what is still needed or not
+// TODO: implemented twice. this moved to unstructured pkg, check what is still needed or not
 const (
 	operationCreate = "create"
 	operationSubmit = "submit"
